git-svn-id: https://swt.informatik.uni-augsburg.de/svn/flashix-codegen/code-btree@541 2e4bb301-4240-4ca3-99d8-184fffe7827c

+++ src/misc/misc.scala
@@ -17,7 +17,7 @@ package object misc {
    * * uninterpreted sorts must be immutable
    */
   
-  type address = Byte
+  type address = Int
   def uninit_address() : address = ???
 
   /**
